requirements.txt
# Twitter API
tweepy>=4.14.0

# Notion API
notion-client>=2.2.0

# Configuration and environment
python-dotenv>=1.0.0

# HTTP requests (backup for manual API calls)
requests>=2.31.0

# Rate limiting
ratelimit>=2.2.1

# Logging
python-json-logger>=2.0.7

# File locking for state management
filelock>=3.13.0
